+++ tests/elasticsearch.ts
@@ -1,21 +1,14 @@
-import {Client} from '@elastic/elasticsearch';
-import {
-  ApiResponse,
-  TransportRequestOptions,
-  TransportRequestPromise
-} from '@elastic/elasticsearch/lib/Transport';
+import {Client, estypes as EsTypes} from '@elastic/elasticsearch';
 
 let client: undefined | Client;
 
-export function search(options: TransportRequestOptions): TransportRequestPromise<ApiResponse> {
+export function search(options: EsTypes.SearchRequest): Promise<EsTypes.SearchResponse> {
   const es = getClient();
 
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-ignore
   return es.search(options);
 }
 
-export async function refreshAllIndexes(): Promise<TransportRequestPromise<ApiResponse>> {
+export async function refreshAllIndexes(): Promise<EsTypes.IndicesRefreshResponse> {
   const es = getClient();
 
   return es.indices.refresh({index: '_all'});